Updates for improved gross checks

+++ ICE/ice_scoring/master/all.csh
@@ -1,32 +1,57 @@
 #!/bin/csh -f
-#SBATCH -J p31_11
-#SBATCH -e p31_11.err
-#SBATCH -o p31_11.out
+#SBATCH -J p6_11
+#SBATCH -e p6_11.err
+#SBATCH -o p6_11.out
 #SBATCH -t 7:55:00
 #SBATCH -q batch
 #SBATCH -A marine-cpu
 #SBATCH -N 1
 #SBATCH --mail-type FAIL
-#SBATCH --mail-user robert.grumbine@noaa.gov
+#SBATCH --mail-user USER@system
 
+#Orion
+#  tbd
+#WCOSS
+#  tbd
+
+#Hera:
 source /etc/profile.d/modules.csh
-setenv PATH /scratch2/NCEPDEV/climate/Robert.Grumbine/anaconda3/bin:$PATH
+module load intel/2020.2
+module load impi/2020.2
+module load netcdf/4.7.0
+module load wgrib2/2.0.8
 
-setenv expt bm31
-setenv FCST_BASE /scratch2/NCEPDEV/climate/Lydia.B.Stefanova/Models/ufs_b31/SeaIce/ 
-setenv EXDIR /scratch2/NCEPDEV/climate/Robert.Grumbine/${expt}.verf
-cd /scratch2/NCEPDEV/climate/Robert.Grumbine/${expt}.verf
+module use -a /contrib/anaconda/modulefiles
+module load anaconda/latest 
+
+#All systems:
+module list
+
+setenv USER $user
+
+setenv expt p6.0
+setenv FCST_BASE /scratch2/NCEPDEV/climate/Lydia.B.Stefanova/Models/ufs_p6/SeaIce/ 
+setenv EXDIR   /scratch2/NCEPDEV/climate/${USER}/prototype_evaluations/${expt}.verf
+setenv RUNBASE /scratch2/NCEPDEV/stmp1/${USER}/prototype_evaluations/${expt}.verf
+cd /scratch2/NCEPDEV/climate/${USER}/prototype_evaluations/${expt}.verf
 setenv base `pwd`
-setenv RUNBASE /scratch2/NCEPDEV/stmp1/Robert.Grumbine/${expt}.verf
+
+setenv mmablib /scratch2/NCEPDEV/climate/Robert.Grumbine/mmablib/
 
 echo env $FCST_BASE $EXDIR $base $RUNBASE
 
+# Fewer changes below here -------------------------------------------------
+
+#setenv PATH /scratch2/NCEPDEV/climate/Robert.Grumbine/anaconda3/bin:$PATH
+
 #For batch python graphics
-setenv XDG_RUNTIME_DIR /scratch2/NCEPDEV/climate/Robert.Grumbine/runtime
+setenv XDG_RUNTIME_DIR /scratch2/NCEPDEV/climate/${USER}/runtime
 if ( ! -d $XDG_RUNTIME_DIR ) then
   mkdir -p -m 700 $XDG_RUNTIME_DIR
 endif
 echo $XDG_RUNTIME_DIR for python graphic support
+setenv MPLCONFIGDIR /scratch2/NCEPDEV/climate/${USER}/runtime
+
 
 setenv x `date`
 echo start of loop at dtime $x
@@ -58,19 +83,20 @@ foreach yy ( 2011 )
         setenv i 0
         while ( $i < 35 )
           setenv tag `expr $tag + 1`
-          setenv tag `dtgfix3 $tag`
+          setenv tag `$mmablib/ush/dtgfix3 $tag`
           setenv i   `expr $i + 1`
           # -m cProfile -o pystats.$tag.$i to generate profiling 
           #   stats for later analysis. Optional
           python3  $EXDIR/setup_verf_ice.py $initial $tag ${FCST_BASE}/${initial}/6hrly/ >>& $expt.$initial.$tag.out
-          mv fcst_edge.$tag          $OUT
-          mv edge.fcst.*.*.$tag      $OUT
+          if ( -f fcst_edge.$tag ) then
+            mv fcst_edge.$tag          $OUT
+            mv edge.fcst.*.*.$tag      $OUT
+          endif
         end
         mv $expt.$initial.*.out $OUT
 
 # Per-forecast statistics, graphics
-        time python3 $EXDIR/contingency_plots.py 35 $initial 0.15 
-        echo time python3 $EXDIR/contingency_plots.py 35 $initial 0.15 >> $OUT/to_run
+        time python3 $EXDIR/contingency_plots.py 35 $initial 0.15 $expt 
         mv *${initial}.png                $OUT
         mv score*${initial}.csv           $OUT
       else
